changed float value of player energy decrease over time

+++ Assets/Scripts/PlayerData.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 [CreateAssetMenu]
 public class PlayerData : ScriptableObject
 {
-    public int playerEnergy { get; private set; } = 100;
+    public float playerEnergy { get; private set; } = 100;
 
-    public void ChangeEnergy(int data)
+    public void ChangeEnergy(float data)
     {
         playerEnergy += data;
     }
 
-    public void SetEnergy(int data)
+    public void SetEnergy(float data)
     {
         playerEnergy = data;
     }
